requirements.txt
alembic==1.13.3
asttokens==2.4.1
autoflake==2.3.1
category-encoders==2.6.4
certifi==2024.8.30
charset-normalizer==3.4.0
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
debugpy==1.8.7
decorator==5.1.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
factor_analyzer==0.5.1
fonttools==4.54.1
greenlet==3.1.1
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
llvmlite==0.43.0
Mako==1.3.6
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numba==0.60.0
numpy==2.0.2
opencv-python==4.10.0.84
openpyxl==3.1.5
optuna==4.0.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
polars==1.12.0
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyflakes==3.2.0
Pygments==2.18.0
pymrmre==1.0.6
pynndescent==0.5.13
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
radMLBench==1.1
requests==2.32.3
scikit-learn==1.5.2
scikit-posthocs==0.10.0
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
SQLAlchemy==2.0.36
stack-data==0.6.3
statsmodels==0.14.4
threadpoolctl==3.5.0
tomli==2.1.0
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
urllib3==2.2.3
wcwidth==0.2.13
